Neighborly_APP/NeighborlyAPI/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
pymongo==3.10.1